requirements.txt
fastapi[standard]
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
alembic
qdrant-client
google-genai
pymupdf
python-docx
bcrypt
slowapi
python-multipart
pydantic-settings
passlib[argon2]
python-jose[cryptography]
pyotp
qrcode[pil]
redis[asyncio]
fastapi-mail
aiosmtplib
pytest
pytest-asyncio
httpx
pytest
pytest-asyncio
httpx
